refactor(systems-grid): extract theme/background asset handling into a part file
Move the theme/background asset slice of _SystemCardGridViewState
(_onSecondaryStateChanged, _loadThemeAssetsForSystems,
_precacheSystemBackgrounds, _updateSecondaryScreenName) verbatim into a new
part file, my_systems_grid/theme_background.dart, as `extension
_ThemeBackground`.

Pure source-split, no behaviour or public-API change. Moved bodies are
byte-identical to main (token-stream verified). All state stays on the host
State; the 2 setState calls route through the host `rebuild()` bridge
(State.setState is @protected). Host 1303 -> 1108; new file 210 lines.

Local gate: dart format clean, flutter analyze "No issues found", 207/207.
